Fixed bug when user leaves server

+++ src/main/java/main/Commands/AdministrativeCommand.java
@@ -17,6 +17,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.Executors;
@@ -65,8 +66,14 @@ public class AdministrativeCommand extends ConcurrentCommand {
 		System.out.println("USER LEAVED");
 
 		Executors.newSingleThreadExecutor()
-				.execute(() ->
-						getDao().remove(event.getMember().getIdLong())
+				.execute(() -> {
+							getDao().removeFromGuild(event.getMember().getIdLong(), event.getGuild().getIdLong());
+							MultiValueMap<Long, Long> map = getDao().getMapGuildUsers();
+							Collection<Long> list = map.getCollection(event.getMember().getIdLong());
+							if (list == null || list.isEmpty()) {
+								getDao().remove(event.getMember().getIdLong());
+							}
+						}
 				);
 	}
